% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibsearch-package.R
\docType{package}
\name{bibsearch-package}
\alias{bibsearch}
\alias{bibsearch-package}
\title{bibsearch: Facilitate Programmatically Creation of Shareable Files
    of Bibliographic Searches Results}
\description{
Provide simple to use wrappers to API of multiple
    bibliographic search engines and database (e.g., Scopus, PubMed), with
    the specific aim of create shareable files (e.g., bibTeX, RIS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/bibsearch}
  \item Report bugs at \url{https://github.com/CorradoLanera/bibsearch/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

}
\keyword{internal}
