% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bibliography.R
\name{create_bibliography}
\alias{create_bibliography}
\alias{create_bibliography.default}
\alias{create_bibliography.pubmed}
\title{Create bibliography}
\usage{
create_bibliography(retrieved_object)

\method{create_bibliography}{default}(retrieved_object)

\method{create_bibliography}{pubmed}(retrieved_object)
}
\arguments{
\item{retrieved_object}{Object output from a search. At the moment,
the only object managed by \link{create_bibliography} are \code{pubmed},
and \code{scopus} (see \strong{See also} section).}
}
\value{
An object of class \code{\link[revtools]{bibliography-class}}
}
\description{
This function creates  \code{\link[revtools]{bibliography-class}}
object. Methods are defined to manage sources of different types.
}
\seealso{
\link{search_on_pubmed}, \link{search_on_scopus}
}
