% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medline_from_scopus.R
\name{medline_from_scopus}
\alias{medline_from_scopus}
\title{Convert Scopus search to medline format}
\usage{
medline_from_scopus(x)
}
\arguments{
\item{x}{An object of class \code{scopus}}
}
\value{
an object of class \code{medline}
}
\description{
Convert scopus searches by \link{complete_scopus_search} or
\link{standard_scopus_search} to medline bibliography format.
}
\details{
Methods are provided for both "STANDARD" and "COMPLETE" views on
Scopus.
}
\examples{
\dontrun{
    medline_from_scopus(scopus_search("machine learning", 2))
}
}
