% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Perform a search on PubMed}
\usage{
pubmed_search(search_string, max_to_get = 25, api_key = NULL)
}
\arguments{
\item{search_string}{(character vector of length 1) that is used for
querying PubMed (standard PubMed syntax, see reference for
details).}

\item{max_to_get}{Integer (>= 1): maximum number of records to
retrieve. Note that, if they are not finished, a complete batch
will be always retrieved; hence, more than \code{max_to_get} records
will be retrieved Set it to \code{Inf} to be sure to retrieve all
the possible results.}

\item{api_key}{(character vector of length 1): user-specific API key
to increase the limit of queries per second above three. You can
obtain your key from NCBI (see
\url{https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/}}
}
\value{
an object of class \code{pubmed}
}
\description{
This function wraps
\code{easyPubMed::\link[easyPubMed]{get_pubmed_ids}} and
\code{easyPubMed::\link[easyPubMed]{fetch_pubmed_data}} to collect
results of a query to
\href{https://www.ncbi.nlm.nih.gov/pubmed/}{PubMed}
}
\examples{
\dontrun{
    pubmed_search("machine learning", 1)
    pubmed_search("aaaa", 1) # no result
}
}
