% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_searches.R
\name{scopus_search}
\alias{scopus_search}
\alias{complete_scopus_search}
\alias{standard_scopus_search}
\title{Scopus search}
\usage{
scopus_search(search_string, max_to_get = 25)

complete_scopus_search(search_string, max_to_get = 25)

standard_scopus_search(search_string, max_to_get = 25)
}
\arguments{
\item{search_string}{(character vector of length 1) that is used for
querying PubMed (standard Scopus syntax, see reference for
details).}

\item{max_to_get}{Integer (>= 1): maximum number of records to
retrieve. Maximum 20000/week overall!}
}
\value{
an object of class \code{scopus}

an object of class \code{scopus-complete}

an object of class \code{scopus-standard}
}
\description{
Scopus search

Complete Scopus search

Standard scopus search
}
\section{Functions}{
\itemize{
\item \code{complete_scopus_search}: Complete Search

\item \code{standard_scopus_search}: Complete Search
}}

\examples{
\dontrun{
    scopus_search("machine learning", 2)
}
\dontrun{
    complete_scopus_search("machine learning", 2)
}
\dontrun{
    standard_scopus_search("machine learning", 2)
}
}
\seealso{
\link{create_bibliography}.
}
