% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_on_scopus.R
\name{search_on_scopus}
\alias{search_on_scopus}
\title{Perform a search on Scopus}
\usage{
search_on_scopus(search_string, max_to_get = 25)
}
\arguments{
\item{search_string}{(character vector of length 1) that is used for
querying PubMed (standard Scopus syntax, see reference for
details).}

\item{max_to_get}{Integer (>= 1): maximum number of records to
retrieve. Maximum 20000/week overall!}
}
\value{
An object of class
\code{\link[revtools]{bibliography-class}}.
}
\description{
This function wraps
\code{rscopus::\link[rscopus]{scopus_search}} to collect
results of a query to
\href{https://www.scopus.com/}{Scopus}
}
\examples{
\dontrun{
    # This retrieve a (short) set of results (batch size auto-resize
    # itself).
    search_on_scopus("ALL(machine AND learning) AND AUTHOR-NAME (lanera)")
}
}
\references{
\url{http://schema.elsevier.com/dtds/document/bkapi/search/SCOPUSSearchTips.htm}

\url{https://dev.elsevier.com/api_key_settings.html}

\url{https://dev.elsevier.com/tecdoc_developer_faq.html}
}
\seealso{
\link{standard_scopus_search}, \link{complete_scopus_search},
\link{create_bibliography}.
}
