% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_searches.R
\name{standard_scopus_search}
\alias{standard_scopus_search}
\title{Standard scopus search}
\usage{
standard_scopus_search(search_string, max_to_get = 25)
}
\arguments{
\item{search_string}{(character vector of length 1) that is used for
querying PubMed (standard Scopus syntax, see reference for
details).}

\item{max_to_get}{Integer (>= 1): maximum number of records to
retrieve. Maximum 20000/week overall!}
}
\value{
an object of class \code{scopus-standard}
}
\description{
Standard scopus search
}
\seealso{
\link{complete_scopus_search}, \link{search_on_scopus},
\link{create_bibliography}.
}
