% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bibliography.R
\name{write_bibliography}
\alias{write_bibliography}
\title{Export Research form a bibliography}
\usage{
write_bibliography(
  bib,
  file_type = c("ris", "bib"),
  out_dir = here::here(),
  base_name = "exported",
  max_each_file = length(bib)
)
}
\arguments{
\item{bib}{an object of class
\code{\link[revtools]{bibliography-class}}}

\item{file_type}{(character vector of length 1) Format of the
exported file. Should be either "ris" (default) or "bib"}

\item{out_dir}{(character vector of length 1): path for the directory
in which to store the output (default is the current/project one)}

\item{base_name}{(character vector of length 1): base name (w/o
extension) for the output file(s)}

\item{max_each_file}{Integer (>=1) number of record to store
and write into each file. Default is \code{length(bib)}, which means
that a single file will be written including all the records in the
bibliography.}
}
\value{
invisibly \code{TRUE}.
}
\description{
This function store in a file a bibliography of class
\code{\link[revtools]{bibliography-class}} both in RIS and BibTeX
file-format.,Moreover, it includes an option to split the
bibliography in multiple files accordingly to an user-defined number
of records to write in each file.
}
\examples{
\dontrun{
    # This save the (short) set of results in the current project
    # directory as a RIS file.
    biblio_short <- search_on_pubmed(
        "(machine learning) AND lanera[Author]"
    )
    write_bibliography(biblio_short)
    write_bibliography(biblio_short, file_type = "bib")
#'
    # Big ammount of reults here (we retain only a few).
    ml_reduced <- search_on_pubmed("machine learning",
        max_to_get = 50
    )

    # Write some RIS files, each with a limited pre-defined ammount
    # of record, and save them in a specific folder.
    write_bibliography(ml_reduced,
        out_dir = "multiple",
        max_each_file = 25
    )

    # Or save them all into a single file (default)
    write_bibliography(ml_reduced, out_dir = "single")
}
}
