% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bibliography.R
\name{create_bibliography}
\alias{create_bibliography}
\alias{create_bibliography.default}
\alias{create_bibliography.medline}
\alias{create_bibliography.scopus}
\title{Create bibliography}
\usage{
create_bibliography(retrieved_object)

\method{create_bibliography}{default}(retrieved_object)

\method{create_bibliography}{medline}(retrieved_object)

\method{create_bibliography}{scopus}(retrieved_object)
}
\arguments{
\item{retrieved_object}{Object output from a search. At the moment,
the only object managed by \link{create_bibliography} is \code{medline},
as produced by \link{search_on_pubmed} or \link{search_on_scopus} .}
}
\value{
An object of class \code{\link[revtools]{bibliography-class}}
}
\description{
This function creates  \code{\link[revtools]{bibliography-class}}
object. Methods are defined to manage sources of different types.
}
\examples{
\dontrun{
    create_bibliography(pubmed_search("machine learning", 2))
    create_bibliography(scopus_search("machine learning", 2))
}
}
\seealso{
\link{search_on_pubmed}, \link{search_on_scopus}
}
